% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_read.R
\name{fars_map_state}
\alias{fars_map_state}
\title{Map of accidents}
\usage{
fars_map_state(state.num, year)
}
\arguments{
\item{state.num}{An integer giving the state code}

\item{year}{A character string/integer giving the year of the input data}
}
\description{
Visualization of a state map showing places (coordinates) of accidents in a specific year
}
\details{
If the state number "state.num" is invalid "invalid STATE number: <state number>" is returned

If there are no accidents for the state in this year "no accidents to plot" is returned

Longitude (latitude) in plot is limited to 900 (90)
}
\examples{
fars_map_state(13,2014)

}
