% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_read.R
\name{fars_read}
\alias{fars_read}
\title{Reading and converting in FARS data into data frame}
\usage{
fars_read(filename)
}
\arguments{
\item{filename}{A character string giving the name of the file}
}
\value{
A data frame read from the csv-file, if file doesnt exist an error
}
\description{
The function reads a csv-file with data from the US National Highway Traffic Safety
Administration's Fatality Analysis Reporting System (FARS) into a tibble data frame
FARS is a nationwide census providing the American public yearly data
(available data: 2013 - 2015) regarding fatal injuries suffered in motor vehicle traffic crashes
}
\examples{
ex_year <- 2013
data <- ex_year \%>\%
  make_filename \%>\%
  fars_read
head(data)

}
