% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_read.R
\name{fars_read_years}
\alias{fars_read_years}
\title{Reading of FARS years and months}
\usage{
fars_read_years(years)
}
\arguments{
\item{years}{A vector (of integers) giving a list of years}
}
\value{
A data frame including values in data by month or NULL if the year is invalid
}
\description{
Reading of FARS years and months
}
\examples{
fars_read_years(c(2013, 2014))

}
