% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_read.R
\name{fars_summarize_years}
\alias{fars_summarize_years}
\title{Summary of FARS data}
\usage{
fars_summarize_years(years)
}
\arguments{
\item{years}{A vector (of integers) giving a list of years to summarize}
}
\value{
A tibble data frame including number of accidents summarized by month and year
}
\description{
The function summarizes accidents data by month and year over multiple input files
}
\examples{
fars_summarize_years(c(2013, 2014))

}
