% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/farspack_package.R
\docType{package}
\name{farspack}
\alias{farspack}
\alias{farspack-package}
\title{Analysis of FARS data}
\description{
Contains five function for reading flat files,
summarizing and visualizing annual occurance of trafic accidents
for each state.
}
\author{
Elena Alexandrova \email{helex07@gmail.com}
}
