% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_read.R
\name{make_filename}
\alias{make_filename}
\title{Creation of csv-file name}
\usage{
make_filename(year)
}
\arguments{
\item{year}{A character string/integer giving the year of the input data}
}
\value{
A string with the data file name for a specific year
}
\description{
The function needs a year as input and produces a valid FARS filename.
It will not check whether the file actually exists, or if the year is reasonable.
}
\examples{
make_filename(2013)

}
