% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_map_state}
\alias{fars_map_state}
\title{Creates state map with accident of given year}
\usage{
fars_map_state(state.num, year)
}
\arguments{
\item{state.num}{This is a number corresponding to the state of which the map will be created}

\item{year}{This is the years of the data that will be presented}
}
\value{
This function returns the mapped data of the respective years for the states of interest
}
\description{
This is a function that, takes a sate and set of yeyars as input,
the data is plotted ina a mapof the state for the respective years
}
\note{
An error will occur if a file does not exist for those yerass or the years inout are not valid
or if the numbers do not correspond to a state
}
\examples{
\dontrun{
fars_map_state(1,2013)
}

}
