% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_map_state.R
\name{fars_map_state}
\alias{fars_map_state}
\title{This function takes year and a state number and plots the map of accidents. if the state number does not exist
it will return an error. and if there are no accidents in a state it will stop and returns a message}
\usage{
fars_map_state(state.num, year)
}
\arguments{
\item{state.num}{number of the target state}

\item{year}{The target year}
}
\description{
#' @references
  \url{https://www.nhtsa.gov/research-data/fatality-analysis-reporting-system-fars}
}
