% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_map_state.R
\name{fars_map_state}
\alias{fars_map_state}
\title{This function takes year and a state number and plots the map of accidents. if the state number does not exist
it will return an error. and if there are no accidents in a state it will stop and returns a message}
\usage{
fars_map_state(state.num, year)
}
\arguments{
\item{state.num}{number of the target state}

\item{year}{The target year}
}
\value{
a plot of accidents in the given state and the given year. it returns an error if the state does not exist and returns a message if there are no accidents in the given state
}
\description{
This function takes year and a state number and plots the map of accidents. if the state number does not exist
it will return an error. and if there are no accidents in a state it will stop and returns a message
}
\examples{
fars_map_state(1, 2013)
}
