% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/farspack_functions.R
\name{fars_read}
\alias{fars_read}
\title{Read .csv file}
\usage{
fars_read(filename)
}
\arguments{
\item{filename}{The name of data file as a character string.}
}
\value{
The data file as a data frame.
}
\description{
The function loads US FARS data \url{https://www.nhtsa.gov/research-data/fatality-
analysis-reporting-system-fars}  and turns flat file into a data frame. If file doesn't
exist, a warning message will appear.
}
\examples{
\dontrun{
data_2013 <- fars_read(filename = "accident_2013.csv.bz2")
}

}
