% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_read}
\alias{fars_read}
\title{Read csv file}
\usage{
fars_read(filename)
}
\arguments{
\item{filename}{Name of the csv file that will be read as table data}
}
\value{
This function returns a table using the data in the \code{filename} file
the phrase.
}
\description{
This is a simple function that, by default, reads the file specified with the
\code{filename} argument.
}
\note{
An error will occur if the file does not exist
}
\examples{
\dontrun{
fars_read('accident_2013.csv.bz2')
}

}
