% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_read.R
\name{fars_read}
\alias{fars_read}
\title{reads a file into a tibble
if the file doesn't exist it returns an error "The file does not exist"}
\usage{
fars_read(filename)
}
\arguments{
\item{filename}{name of the file}
}
\value{
returns a tbl_df if the file exists in the directory and will return an error otherwise
}
\description{
reads a file into a tibble
if the file doesn't exist it returns an error "The file does not exist"
}
\examples{
fars_read('accident_2013.csv.bz2')
}
