% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/farspack_functions.R
\name{fars_read_years}
\alias{fars_read_years}
\title{Create a data file for a specified date of accidents.}
\usage{
fars_read_years(years)
}
\arguments{
\item{years}{The years of the accidents as a list or vector of integers}
}
\value{
The accidents for given years sorted by month as list of data frames with
integers and numeric elements. Returns NULL if entered year doesn't exist in the data
set.
}
\description{
The function creates a subset that contains accidents occured in the particular years
and sorted by months. A warning message is issued if the specified year is not in the
original dataset.
}
\examples{
\dontrun{
fars_read_years(years = 2013:2015)
fars_read_years(years = c(2013, 2014, 2015))
fars_read_years(years = list(2013 2014, 2015))
}

}
