% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_read_years}
\alias{fars_read_years}
\title{Read accident files}
\usage{
fars_read_years(years)
}
\arguments{
\item{years}{This is the years of the data that will be hold in the files}
}
\value{
This function returns the data of the respective years
}
\description{
This is a simple function that, takes an array of years as input,
to read the files that hold the accidents for those repsectcive years
}
\note{
An error will occur if a file does not exist for those yerass or the years inout are not valid
}
\examples{
fars_read_years(c(2013,2014))

}
