% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_read_years.R
\name{fars_read_years}
\alias{fars_read_years}
\title{this function takes a vector of years and opens the file related to each year and imports it into a tbl_df
it then takes the part of the tbl_df that belongs to that specific year. then it  takes columns MONTH and year of that tbl_df
and returns it in the list
if the file does not exist it will return a warning and returns a NULL inside the list}
\usage{
fars_read_years(years)
}
\arguments{
\item{year}{a vector of integers where each integer is a year,}
}
\value{
a list of tbl_df each having two columns MONTH and year and NULL for the non-existing Years with a warning
}
\description{
this function takes a vector of years and opens the file related to each year and imports it into a tbl_df
it then takes the part of the tbl_df that belongs to that specific year. then it  takes columns MONTH and year of that tbl_df
and returns it in the list
if the file does not exist it will return a warning and returns a NULL inside the list
}
\examples{
tibble_vector <- fars_read_years(c(2000, 2001))
}
