% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_read_years.R
\name{fars_read_years}
\alias{fars_read_years}
\title{this function takes a vector of years and opens the file related to each year and imports it into a tbl_df
it then takes the part of the tbl_df that belongs to that specific year. then it  takes columns MONTH and year of that tbl_df
and returns it in the list
if the file does not exist it will return a warning and returns a NULL inside the list}
\usage{
fars_read_years(years)
}
\arguments{
\item{years}{target years}
}
\description{
#' @references
  \url{https://www.nhtsa.gov/research-data/fatality-analysis-reporting-system-fars}
}
\examples{

\dontrun{
fars_read_years(c(2016,2017))
}

}
