% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_summarize_years}
\alias{fars_summarize_years}
\title{Summarize FARS crash data by year and month}
\usage{
fars_summarize_years(years)
}
\arguments{
\item{years}{Integer vector specifying year(s) to summarize.}
}
\value{
A data frame with counts of fatal crashes by year and month.
}
\description{
Summarize FARS crash data by year and month
}

