% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/farspack_functions.R
\name{fars_summarize_years}
\alias{fars_summarize_years}
\title{Calculate number of accidents per month for the particular years.}
\usage{
fars_summarize_years(years)
}
\arguments{
\item{years}{The years of the accidents as a list or vector of integers}
}
\value{
The number of the accidents per month for each particular year as a data frame.
}
\description{
The function is summarising  the accidents for the selected years and counts the number
of reports per month.
}
\examples{
\dontrun{
fars_summarize_years(2013:2015)
fars_summarize_years(c(2013, 2014, 2015))
fars_summarize_years(list(2013, 2014, 2015))
}

}
