% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_summarize_years}
\alias{fars_summarize_years}
\title{Read and summarize accident files}
\usage{
fars_summarize_years(years)
}
\arguments{
\item{years}{This is the years of the data that will be summarized}
}
\value{
This function returns the data of the respective years
}
\description{
This is a simple function that, takes an array of years as input,
to read the files that hold the accidents for those repsectcive years and summarize the data
}
\note{
An error will occur if a file does not exist for those yerass or the years inout are not valid
}
\examples{
\dontrun{
fars_summarize_years(c(2013,2014))
}

}
