% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_summerize_years.R
\name{fars_summarize_years}
\alias{fars_summarize_years}
\title{this function takes a vector of years. reads the file for each year. creates a list of tbl_df, it does a row binding over all the tbls
it groups the resulting tibble by year and MONTH and then counts the occurances and then it makes a wide tibble from the long tibble according to year.}
\usage{
fars_summarize_years(years)
}
\arguments{
\item{years}{The target years}
}
\value{
returns a spreaded tbl_df if the file exists and is read into R successfully and NULL if the file does not exist.
}
\description{
#' @references
  \url{https://www.nhtsa.gov/research-data/fatality-analysis-reporting-system-fars}
}
\examples{
\dontrun{
fars_summerized_year(c(2013, 2014))
}

}
