% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/farspack_functions.R
\name{make_filename}
\alias{make_filename}
\title{Give name to a data file}
\usage{
make_filename(year)
}
\arguments{
\item{year}{The year of the accidents as an integer.}
}
\value{
The file name as a character string.
}
\description{
The function creates a file name for the specific year and prints it.
The file name contains the year when the accidents occured.
}
\examples{
\dontrun{
make_filename(year = 2013)
}

}
