% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{make_filename}
\alias{make_filename}
\title{Create file name}
\usage{
make_filename(year)
}
\arguments{
\item{year}{This is the year of the data that will be hold in the file}
}
\value{
This function returns the proper name for the file containing the accidents data
from the year specified as an argument.
}
\description{
This is a simple function that, takes a string argument, \code{year},
to create the name of the file that will hold the accidents for that repsectcive year
}
\note{
An error will occur if not a  proper year is specicfied as an argument
}
\examples{
\dontrun{
make_filename(2012)
}
}
