% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_filename.R
\name{make_filename}
\alias{make_filename}
\title{generates the name of the file according to the year}
\usage{
make_filename(year)
}
\arguments{
\item{year}{a string or an integer to address the file that we want to read}
}
\value{
string which is going to be name of the
}
\description{
generates the name of the file according to the year
}
\examples{
fileName <- make_file(2014)
}
